make tbp interface work on web hopefully
